feat: Finish day 10 part 2
To solve part two I took the pipe map (an ordered array of segments)
built by part one, and traveled it. Keeping track of every left
and right turn I was able to determine if i was going Clockwise or
Counter Clockwise around the loop. At each section of pipe I keep
track of what tiles are to the left and right for the pipe (inculding
before and after turns). If the loop is Clockwise then Right side
tiles are inside, and for Counter clockwise left side tiles are inside

Finally, I then took all my inside tiles, and for every tile, I checked
if its neighbors were a Part of the pipe, or already accounted for
If not I added it as an inside tile, and then did this again for that
tile. This way I was able to fill in entire contained areas.